Adds authentication on the backend

+++ backend/businessLogic/entriesResolver.ts
@@ -1,11 +1,12 @@
-// import { createLogger } from "../utils/logger";
-// const logger = createLogger("resolver");
-
 export const resolvers = {
   Query: {
-    getEntries: async (_, args, { dataSources }) => {
+    getEntries: async (_, args, { dataSources }, context) => {
       try {
-        return await dataSources.entriesAccess.getEntries(args.userId);
+        if (context.user === args.userId) {
+          return await dataSources.entriesAccess.getEntries(args.userId);
+        } else {
+          console.log("User is not authorized to perform this action");
+        }
       } catch (error) {
         console.error(error);
       }
@@ -13,33 +14,45 @@ export const resolvers = {
   },
 
   Mutation: {
-    createEntry: async (_, args, { dataSources }) => {
+    createEntry: async (_, args, { dataSources }, context) => {
       try {
-        return await dataSources.entriesAccess.createEntry(
-          args.userId,
-          args.entryInput
-        );
+        if (context.user === args.userId) {
+          return await dataSources.entriesAccess.createEntry(
+            args.userId,
+            args.entryInput
+          );
+        } else {
+          console.log("User is not authorized to perform this action");
+        }
       } catch (error) {
         console.error(error);
       }
     },
-    updateEntry: async (_, args, { dataSources }) => {
+    updateEntry: async (_, args, { dataSources }, context) => {
       try {
-        return await dataSources.entriesAccess.updateEntry(
-          args.userId,
-          args.entryId,
-          args.entryInput
-        );
+        if (context.user === args.userId) {
+          return await dataSources.entriesAccess.updateEntry(
+            args.userId,
+            args.entryId,
+            args.entryInput
+          );
+        } else {
+          console.log("User is not authorized to perform this action");
+        }
       } catch (error) {
         console.error(error);
       }
     },
-    deleteEntry: async (_, args, { dataSources }) => {
+    deleteEntry: async (_, args, { dataSources }, context) => {
       try {
-        return await dataSources.entriesAccess.deleteEntry(
-          args.userId,
-          args.entryId
-        );
+        if (context.user === args.userId) {
+          return await dataSources.entriesAccess.deleteEntry(
+            args.userId,
+            args.entryId
+          );
+        } else {
+          console.log("User is not authorized to perform this action");
+        }
       } catch (error) {
         console.error(error);
       }
